Change int to float

+++ db/migrate/20160623204625_create_kml_settlements.rb
@@ -2,8 +2,8 @@ class CreateKmlSettlements < ActiveRecord::Migration
   def change
     create_table :kml_settlements do |t|
       t.string :name
-      t.integer :geo_x
-      t.integer :geo_y
+      t.float :geo_x
+      t.float :geo_y
       t.integer :settlement_id
     end
   end
